+++ app/src/pages/ArticlePage.js
@@ -1,7 +1,7 @@
-import React from 'react'
-import Section from "./ArticleSection"
-import ArticleHeader from "./ArticleHeader"
-import {getFullArticle} from '../utils/functions/articles'
+import React from 'react';
+import Section from "../components/Articles/ArticleSection";
+import ArticleHeader from "../components/Articles/ArticleHeader";
+import { getFullArticle } from '../utils/functions/articles';
 
 class ArticlePage extends React.Component {
     constructor(props) {
@@ -18,22 +18,22 @@ class ArticlePage extends React.Component {
                 article: article
             });
         }).catch((err) => {
-            window.location.href=('/article-not-found');
+            window.location.href = ('/article-not-found');
         })
     }
 
     render() {
         // If article hasn't loaded yet, render div
         // TODO: render loading animation here
-        if(!this.state.article) {
-            return(
+        if (!this.state.article) {
+            return (
                 <div>
 
                 </div>
             )
         }
         // Render the article
-        return(
+        return (
             <div>
                 <ArticleHeader article={this.state.article} />
                 {this.state.article.sections.map((section) =>
